Feedback bettered and attribute percentage of Coupon changed into float

+++ couponProject/couponApp/models.py
@@ -5,4 +5,4 @@ from django.db import models
 class Coupon(models.Model):
 	name         = models.CharField(max_length=100, unique=True)
 	availability = models.IntegerField(default=1)
-	percentage   = models.IntegerField()
+	percentage   = models.FloatField()
